% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bboread.R
\name{bboread}
\alias{bboread}
\title{Read in BBO data sets from CME Historical Datamine}
\usage{
bboread(x)
}
\arguments{
\item{x}{is a path to a .txt file from the CME Group's historical datamine.}
}
\value{
A datatable.
}
\description{
This function is simply a wrapper for the read_fwf function in the readr package. Inside the function, start and end positions
of the raw fixed width file are specified as documented at
\url{http://www.cmegroup.com/confluence/display/EPICSANDBOX/Best+Bid+Best+Offer+ASCII+Layout}
}
\details{
For performance, only the bare minimum columns are read in.
}
\examples{
DATA <- bboread('data-raw/XCBT_C_FUT_110110.TXT')
head(DATA)

DATA <- bboread('C:/Users/mallorym/BBOCORNDATA/2012Jan-2013Nov_txt/BBO_CBT_20120102-20131130_9552_00.txt')  # A 'large' file.
head(DATA)
}

