% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decimalprices.R
\name{decimalprices}
\alias{decimalprices}
\title{Convert eighth of a cent grain prices to decimal}
\usage{
decimalprices(x)
}
\arguments{
\item{x}{is a column of a datatable of Grain contract BBO raw data from CME Group's Datamine.
Formatted as.character() from the start it will have 7 characters. Positions: 4 hundreds, 5 tens, 6 ones, and 7 8th of a cent
since futures quotes are in cents per bushel and the ticks are 8ths of a cent
this function will only work for corn prices < $100/bushel. At this date, grain prices
above $100/ bushel are unimaginable.

This function is intended to be vectorized over a data.table object as in the example below.}
}
\value{
The column of the orginal datatable with the TrPrice column converted to numeric in decimal format.
}
\description{
Convert eighth of a cent grain prices to decimal
}
\examples{
accum <- as.list(Null)
accum[[1]] <- as.data.table(bboread('XCBT_C_FUT_110110.TXT'))
accum[[2]] <- as.data.table(bboread('XCBT_C_FUT_110110.TXT'))                 # accum is a list of two 'days' worth of BBO data
corn_110110 <- data.table::rbindlist(accum)
corn_110110[, Price := decimalprices(corn_110110$TrPrice)]
}

