% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hour.R
\name{hour}
\alias{hour}
\title{Hours}
\usage{
hour(x)
}
\arguments{
\item{x}{a one column of a data.table containing a time in the ITime, "09:30:00" represents hours, minutes, and seconds.}
}
\value{
An integer.
}
\description{
This function takes time in a time object and returns the hour value as an integer. As documented by Hadley
}
\details{
\url{https://gist.github.com/hadley/10238}
}
\examples{
datasample <- as.data.table(bboread('data-raw/XCBT_C_FUT_110110.TXT'))
datasample[, c("TradeDate", "TradeTime") := .(datemanip(datasample[, 1, with=FALSE]), timemanip(datasample[, 2, with=FALSE]))]
datasample
hour(datasample$TradeTime)
}

