% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RollContracts.R
\name{RollContracts}
\alias{RollContracts}
\title{RollContracts}
\usage{
RollContracts(accum, r = 20)
}
\arguments{
\item{r}{an integer that specifies the date in the month prior to expiration you want to roll. 20 is the default.}

\item{x}{a list of data.tables with at least two columns named 'TradeDate' and 'DeliveryDate'}
}
\value{
An list of data.tables with a column called 'Deferreds' that specifies if the contract is the Nearby, 1st Deferred, or 2nd Deferred.
}
\description{
This function rolls contracts into 'Nearby', '1st Deferred', and '2nd Deferred'. It removes contracts
more distant than 2 Deferred. It takes as arguments, x, a list of data.tables and r, an integer
representing the date on which the nearby contract should be rolled. The data.tables  in the list
can contain raw trades and quotes data after reading in with bboread(), or it can be already aggregated in some
way. For example, if you calculate mean price per day and number of quotes as soon as you read in, you can still call
RollContracts() on this list.
}
\details{
All September Maturities are deleted to avoid, new-crop vs old-crop delivery effects.
}
\examples{
accum <- as.list(NULL)
accum[[1]] <- corn_110110
accum[[2]] <- corn_110110                 # accum is a list of two 'days' worth of BBO data
accum <- RollContracts(accum, r = 20)
}

