% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timemanip.R
\name{timemanip}
\alias{timemanip}
\title{Manipulate Time Strings ITime Format}
\usage{
timemanip(x)
}
\arguments{
\item{x}{a one column of a data.table containing a time in the format, "093000", to represent "09:30:00" hours, minutes, and seconds.}
}
\value{
An ITime object.
}
\description{
This function takes time in the format, "093000", to represent hours, minutes, and seconds. The value of the
function returns an ITime object. After reformatting the time to "09:30:00", the function passes the unambiguous string to
an as.ITime() function call.
}
\details{
This function is intended to be vectorized over a data.table object. See examples below.
}
\examples{
datasample <- as.data.table(bboread('data-raw/XCBT_C_FUT_110110.TXT'))
datasample[, TradeTime .= timemanip(datasample[, 2, with=FALSE])]
datasample

datasample <- as.data.table(bboread('data-raw/XCBT_C_FUT_110110.TXT'))
datasample[, c("TradeDate", "TradeTime") := .(datemanip(datasample[, 1, with=FALSE]), timemanip(datasample[, 2, with=FALSE]))]
datasample
}

