\name{corn_110110}
\alias{corn_110110}
\docType{data}
\title{
Best Bid Best Offer Data for Corn Futures Data from the CME Group
}
\description{
This top-of-the book data set for January 1, 2011 is the BBO data sample provided by the CME Datamine historical archive.
}
\usage{data("corn_110110")}
\format{
  A data table.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
This exaple data set has already been processed as an .rda file. The raw data provided by the CME Group is a fixed width file.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}

\keyword{datasets}
