% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datemanip.R
\name{datemanip}
\alias{datemanip}
\title{Manipulate Date Strings IDate Format}
\usage{
datemanip(x)
}
\arguments{
\item{x}{a one column of a data.table containing a date in the format, "20080114", to represent January 14, 2008.}
}
\value{
An IDate object.
}
\description{
This function takes dates that are formatted as a character string of the type, '20080114', where this is
January 14, 2008 The value of the function returns an IDate object.
}
\details{
This function is intended to be vectorized over a data.table object.
}
\examples{
datasample <- as.data.table(bboread('data-raw/XCBT_C_FUT_110110.TXT'))
datasample[, TradeDate := datemanip(datasample[, 1, with=FALSE])]
datasample

datasample <- as.data.table(bboread('data-raw/XCBT_C_FUT_110110.TXT'))
datasample[, c("TradeDate", "TradeTime") := .(datemanip(datasample[, 1, with=FALSE]), timemanip(datasample[, 2, with=FALSE]))]
datasample
}

